# open-api-contracts/api.yaml
openapi: 3.1.0
info:
  title: Blog Engine API
  description: API for managing Astro blog sites
  license:
    name: ''
  version: 0.1.0
paths:
  /healthz:
    get:
      tags: []
      operationId: healthz
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /preview:
    delete:
      tags: []
      operationId: stop_preview
      responses:
        '204':
          description: Preview stopped (or was not running)
  /sites:
    get:
      tags: []
      operationId: list_sites
      responses:
        '200':
          description: List of all sites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteListResponse'
    post:
      tags: []
      operationId: create_site
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteRequest'
        required: true
      responses:
        '201':
          description: Site created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '400':
          description: Invalid request
        '500':
          description: Management API error
  /sites/{slug}:
    delete:
      tags: []
      operationId: delete_site
      parameters:
        - name: slug
          in: path
          description: Site slug
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Site deleted
        '404':
          description: Site not found
        '500':
          description: Dev server failed to start
  /sites/{slug}/preview:
    post:
      tags: []
      operationId: preview_site
      parameters:
        - name: slug
          in: path
          description: Site slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dev server started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '404':
          description: Site not found
        '500':
          description: Dev server failed to start
components:
  schemas:
    CreateSiteRequest:
      type: object
      description: Request body for creating a new site.
      required:
        - name
        - slug
      properties:
        name:
          type: string
        slug:
          type: string
    HealthData:
      type: object
      required:
        - status
        - version
      properties:
        status:
          $ref: '#/components/schemas/HealthStatus'
        version:
          type: string
    HealthResponse:
      type: object
      description: Response envelope for the health endpoint.
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/HealthData'
        meta:
          $ref: '#/components/schemas/Meta'
    HealthStatus:
      type: string
      enum:
        - healthy
    Meta:
      type: object
      required:
        - timestamp
        - requestId
        - version
        - serverName
      properties:
        requestId:
          type: string
        serverName:
          $ref: '#/components/schemas/MetaServerName'
        timestamp:
          type: string
        version:
          type: string
    MetaServerName:
      type: string
      enum:
        - backend
    SiteData:
      type: object
      description: A single Astro site managed by the CMS.
      required:
        - slug
        - name
        - gitUrl
      properties:
        gitUrl:
          type: string
        name:
          type: string
        previewUrl:
          type:
            - string
            - 'null'
        slug:
          type: string
    SiteListResponse:
      type: object
      description: Response envelope for a list of sites.
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SiteData'
        meta:
          $ref: '#/components/schemas/Meta'
    SiteResponse:
      type: object
      description: Response envelope for a single site.
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/SiteData'
        meta:
          $ref: '#/components/schemas/Meta'